Unified text field hint API and added keyboard handlers to SuperTextField
