+++ super_editor/lib/src/infrastructure/super_textfield/infrastructure/hint_text.dart
@@ -1,12 +1,48 @@
 import 'package:flutter/material.dart';
 import 'package:super_editor/src/default_editor/attributions.dart';
 import 'package:super_editor/src/infrastructure/attributed_spans.dart';
+import 'package:super_editor/src/infrastructure/attributed_text.dart';
+
+/// Policy that dictates when to display a hint in a Super Text Field.
+enum HintBehavior {
+  /// Display a hint when the text field is empty until
+  /// the text field receives focus, then hide the hint.
+  displayHintUntilFocus,
+
+  /// Display a hint when the text field is empty until
+  /// at least 1 character is entered into the text field.
+  displayHintUntilTextEntered,
+
+  /// Do not display a hint.
+  noHint,
+}
+
+/// Builds a hint widget based on given [hintText] and a [hintTextStyleBuilder].
+class StyledHintBuilder {
+  StyledHintBuilder({
+    this.hintText,
+    this.hintTextStyleBuilder = defaultHintStyleBuilder,
+  });
+
+  /// Text displayed when the text field has no content.
+  final AttributedText? hintText;
+
+  /// Text style factory that creates styles for the [hintText],
+  /// which is displayed when [textController] is empty.
+  final AttributionStyleBuilder hintTextStyleBuilder;
+
+  Widget build(BuildContext context) {
+    return Text.rich(
+      hintText?.computeTextSpan(hintTextStyleBuilder) ?? TextSpan(text: "", style: hintTextStyleBuilder({})),
+    );
+  }
+}
 
 /// Creates default [TextStyles] for hint text in a super text field.
 TextStyle defaultHintStyleBuilder(Set<Attribution> attributions) {
   TextStyle newStyle = const TextStyle(
     color: Colors.grey,
-    fontSize: 13,
+    fontSize: 16,
     height: 1.4,
   );
