+++ super_editor/lib/src/infrastructure/super_textfield/ios/_caret.dart
@@ -76,8 +76,7 @@ class IOSTextFieldCaret extends StatefulWidget {
   _IOSTextFieldCaretState createState() => _IOSTextFieldCaretState();
 }
 
-class _IOSTextFieldCaretState extends State<IOSTextFieldCaret>
-    with SingleTickerProviderStateMixin {
+class _IOSTextFieldCaretState extends State<IOSTextFieldCaret> with SingleTickerProviderStateMixin {
   late CaretBlinkController _caretBlinkController;
 
   @override
@@ -169,12 +168,11 @@ class _IOSCursorPainter extends CustomPainter {
     caretPaint.color = caretColor.withOpacity(blinkController.opacity);
 
     final textPosition = selection.extent;
-    final caretHeight =
-        textLayout.getCharacterBox(textPosition).toRect().height;
+    double caretHeight = textLayout.getCharacterBox(textPosition).toRect().height;
+    caretHeight = caretHeight > 0 ? caretHeight : textLayout.getHeightForCaret(textPosition) ?? 0;
+    caretHeight = caretHeight > 0 ? caretHeight : textLayout.getLineHeightAtPosition(textPosition);
 
-    Offset caretOffset = isTextEmpty
-        ? Offset.zero
-        : textLayout.getOffsetAtPosition(textPosition);
+    Offset caretOffset = isTextEmpty ? Offset.zero : textLayout.getOffsetAtPosition(textPosition);
 
     if (borderRadius == BorderRadius.zero) {
       canvas.drawRect(
