+++ super_editor/lib/src/infrastructure/super_textfield/ios/ios_textfield.dart
@@ -3,13 +3,14 @@ import 'package:flutter/services.dart';
 import 'package:super_editor/src/default_editor/super_editor.dart';
 import 'package:super_editor/src/infrastructure/_listenable_builder.dart';
 import 'package:super_editor/src/infrastructure/_logging.dart';
+import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/attributed_text_editing_controller.dart';
 import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/hint_text.dart';
 import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/text_scrollview.dart';
+import 'package:super_editor/src/infrastructure/super_textfield/input_method_engine/_ime_text_editing_controller.dart';
 import 'package:super_editor/src/infrastructure/super_textfield/ios/_editing_controls.dart';
 
 import '../../attributed_text.dart';
 import '../../super_selectable_text.dart';
-import '../super_textfield.dart' hide SuperTextFieldScrollview, SuperTextFieldScrollviewState;
 import '_caret.dart';
 import '_floating_cursor.dart';
 import '../../platforms/ios/toolbar.dart';
@@ -30,11 +31,11 @@ class SuperIOSTextField extends StatefulWidget {
     this.textController,
     this.textStyleBuilder = defaultStyleBuilder,
     this.textAlign = TextAlign.left,
-    this.hintText,
-    this.hintTextStyleBuilder = defaultHintStyleBuilder,
+    this.hintBehavior = HintBehavior.displayHintUntilFocus,
+    this.hintBuilder,
     this.minLines,
     this.maxLines = 1,
-    required this.lineHeight,
+    this.lineHeight,
     required this.caretColor,
     required this.selectionColor,
     required this.handlesColor,
@@ -42,7 +43,9 @@ class SuperIOSTextField extends StatefulWidget {
     this.popoverToolbarBuilder = _defaultPopoverToolbarBuilder,
     this.showDebugPaint = false,
     this.onPerformActionPressed,
-  }) : super(key: key);
+  })  : assert(minLines == null || minLines == 1 || lineHeight != null, 'minLines > 1 requires a non-null lineHeight'),
+        assert(maxLines == null || maxLines == 1 || lineHeight != null, 'maxLines > 1 requires a non-null lineHeight'),
+        super(key: key);
 
   /// [FocusNode] attached to this text field.
   final FocusNode? focusNode;
@@ -58,12 +61,13 @@ class SuperIOSTextField extends StatefulWidget {
   /// [textController] based on the attributions in that content.
   final AttributionStyleBuilder textStyleBuilder;
 
-  /// Text displayed when the text field has no content.
-  final AttributedText? hintText;
+  /// Policy for when the hint should be displayed.
+  final HintBehavior hintBehavior;
 
-  /// Text style factory that creates styles for the [hintText],
-  /// which is displayed when [textController] is empty.
-  final AttributionStyleBuilder hintTextStyleBuilder;
+  /// Builder that creates the hint widget, when a hint is displayed.
+  ///
+  /// To easily build a hint with styled text, see [StyledHintBuilder].
+  final WidgetBuilder? hintBuilder;
 
   /// Color of the caret.
   final Color caretColor;
@@ -103,15 +107,16 @@ class SuperIOSTextField extends StatefulWidget {
   ///  * [lineHeight]
   final int? maxLines;
 
-  /// The height of a single line of text in this text field, used
+  /// The height of a single line of text in this text scroll view, used
   /// with [minLines] and [maxLines] to size the text field.
   ///
-  /// An explicit [lineHeight] is required because rich text in this
-  /// text field might have lines of varying height, which would
-  /// result in a constantly changing text field height during scrolling.
-  /// To avoid that situation, a single, explicit [lineHeight] is
-  /// provided and used for all text field height calculations.
-  final double lineHeight;
+  /// An explicit [lineHeight] is required for multi-line text fields
+  /// because rich text in this text scroll view might have lines of
+  /// varying height, which would result in a constantly changing text
+  /// field height during scrolling. To avoid that situation, a single,
+  /// explicit [lineHeight] is provided and used for all text field height
+  /// calculations.
+  final double? lineHeight;
 
   /// The type of action associated with the action button on the mobile
   /// keyboard.
@@ -374,30 +379,19 @@ class _SuperIOSTextFieldState extends State<SuperIOSTextField> with SingleTicker
             child: ListenableBuilder(
               listenable: _textEditingController,
               builder: (context) {
-                final textSpan = _textEditingController.text.text.isNotEmpty
-                    ? _textEditingController.text.computeTextSpan(widget.textStyleBuilder)
-                    : AttributedText(text: 'enter text').computeTextSpan(
-                        (attributions) => widget.textStyleBuilder(attributions).copyWith(color: Colors.grey));
+                final isTextEmpty = _textEditingController.text.text.isEmpty;
+                final showHint = widget.hintBuilder != null &&
+                    ((isTextEmpty && widget.hintBehavior == HintBehavior.displayHintUntilTextEntered) ||
+                        (isTextEmpty &&
+                            !_focusNode.hasFocus &&
+                            widget.hintBehavior == HintBehavior.displayHintUntilFocus));
 
                 return CompositedTransformTarget(
                   link: _textContentLayerLink,
                   child: Stack(
                     children: [
-                      // TODO: switch out textSelectionDecoration and textCaretFactory
-                      //       for backgroundBuilders and foregroundBuilders, respectively
-                      //
-                      //       add the floating cursor as a foreground builder
-                      SuperSelectableText(
-                        key: _textContentKey,
-                        textSpan: textSpan,
-                        textSelection: _textEditingController.selection,
-                        textSelectionDecoration: TextSelectionDecoration(selectionColor: widget.selectionColor),
-                        showCaret: true,
-                        textCaretFactory: IOSTextFieldCaretFactory(
-                          color: _floatingCursorController.isShowingFloatingCursor ? Colors.grey : widget.caretColor,
-                          width: 2,
-                        ),
-                      ),
+                      if (showHint) widget.hintBuilder!(context),
+                      _buildSelectableText(),
                       Positioned(
                         left: 0,
                         top: 0,
@@ -414,6 +408,28 @@ class _SuperIOSTextFieldState extends State<SuperIOSTextField> with SingleTicker
             ),
           ),
         ),
+      ),
+    );
+  }
+
+  Widget _buildSelectableText() {
+    final textSpan = _textEditingController.text.text.isNotEmpty
+        ? _textEditingController.text.computeTextSpan(widget.textStyleBuilder)
+        : AttributedText(text: "").computeTextSpan(widget.textStyleBuilder);
+
+    // TODO: switch out textSelectionDecoration and textCaretFactory
+    //       for backgroundBuilders and foregroundBuilders, respectively
+    //
+    //       add the floating cursor as a foreground builder
+    return SuperSelectableText(
+      key: _textContentKey,
+      textSpan: textSpan,
+      textSelection: _textEditingController.selection,
+      textSelectionDecoration: TextSelectionDecoration(selectionColor: widget.selectionColor),
+      showCaret: true,
+      textCaretFactory: IOSTextFieldCaretFactory(
+        color: _floatingCursorController.isShowingFloatingCursor ? Colors.grey : widget.caretColor,
+        width: 2,
       ),
     );
   }
